New Scene and Car Controller Changes
Added a new scene with the level for the prototype. It includes buildings and red boxes in place of obstacles. I also constrained the Z rotation of the car and added a few lines in the code to keep the Z rotation near 0 after applying force. The Trap script now resets rotation and speed.